Enhance classification training pipeline with hyperparameter tuning and threshold optimization
Key Changes:
- Updated `run_classification_training` to include options for hyperparameter tuning and threshold optimization.
- Integrated Optuna for hyperparameter tuning, improving model performance.
- Added functionality to optimize classification thresholds for better evaluation metrics.
- Enhanced data splitting to include validation sets for improved model training.
- Updated command-line interface to support new options for tuning and optimization.

These enhancements significantly improve the model training process, making it more robust and efficient.

+++ src/features/build_features.py
@@ -317,11 +317,12 @@ class FeatureEngineer:
 
         X = df[feature_cols].copy()
 
-        # Use late_delivery_risk directly if available, otherwise create from delivery_status
+        # Use late_delivery_risk directly if available, otherwise use late_delivery
+        # Always convert to int for consistency and robust model training
         if 'late_delivery_risk' in df.columns:
             y = df['late_delivery_risk'].astype(int)
         elif 'late_delivery' in df.columns:
-            y = df['late_delivery']
+            y = df['late_delivery'].astype(int)
         else:
             y = None
